refactor: move queries under `query` dir
