+++ src/pages/human-rights/mockQueries.ts
@@ -6,9 +6,9 @@ import {
   HumanRightsPostResponse,
   HumanRightsPostSummaryResponse,
 } from '@/pages/human-rights/schema.ts';
-import { PostsResponse } from '@/pages/human-rights/hooks/useSearchPosts.ts';
+import { PostsResponse } from '@/pages/human-rights/hooks/query/useSearchPosts.ts';
 import { ApiResponse } from '@/pages/human-rights/hooks/useStuQuery.ts';
-import { GetCommentsResponse } from '@/pages/human-rights/hooks/useGetComments.ts';
+import { GetCommentsResponse } from '@/pages/human-rights/hooks/query/useGetComments.ts';
 
 async function waitSecondAndReturn<T>(ms: number, data: T): Promise<ApiResponse<T>> {
   await new Promise<void>((resolve) =>
